# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SurveyAI API
  version: 1.0.0
  description: OpenAPI specification for the current SurveyAI backend endpoints.
servers:
  - url: http://localhost:8080
    description: Local Docker/Desktop backend
...
paths:
...
  /api/v1/operations:
    get:
      tags: [Operations]
      summary: List operations by company
      parameters:
        - $ref: "#/components/parameters/CompanyIdQuery"
      responses:
        "200":
          description: Operations returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OperationResponse"
        default:
          $ref: "#/components/responses/ApiError"
    post:
      tags: [Operations]
      summary: Create operation
      parameters:
        - $ref: "#/components/parameters/CompanyIdQuery"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOperationRequest"
      responses:
        "201":
          description: Operation created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResponse"
        default:
          $ref: "#/components/responses/ApiError"
  /api/v1/operations/{operationId}:
    get:
      tags: [Operations]
      summary: Get operation by id
      parameters:
        - $ref: "#/components/parameters/OperationIdPath"
        - $ref: "#/components/parameters/CompanyIdQuery"
      responses:
        "200":
          description: Operation returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResponse"
        default:
          $ref: "#/components/responses/ApiError"
...
components:
  parameters:
...
    CompanyIdQuery:
      name: companyId
      in: query
      required: true
      schema:
        type: string
        format: uuid
...
    OperationIdPath:
      name: operationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    ApiError:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
  schemas:
...
    CreateOperationRequest:
      type: object
      required:
        - name
        - surveyId
      properties:
        name:
          type: string
          maxLength: 255
        surveyId:
          type: string
          format: uuid
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        createdByUserId:
          type: string
          format: uuid
          nullable: true
    OperationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        companyId:
          type: string
          format: uuid
        surveyId:
          type: string
          format: uuid
        name:
          type: string
        status:
          $ref: "#/components/schemas/OperationStatus"
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdByUserId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ApiErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
...
    OperationStatus:
      type: string
      enum:
        - DRAFT
        - SCHEDULED
        - RUNNING
        - PAUSED
        - COMPLETED
        - CANCELLED
...